+++ CodeWars/Algorithms/BackwardsReadPrimes/BackwardsReadPrimes.rs
@@ -1,8 +1,31 @@
 
+fn is_prime(n: u64) -> bool {
+    if n <= 3 {
+        return true;
+    }
+
+    if n%2 == 0 || n%3 == 0 {
+        return false;
+    }
+
+    let mut i: u64 = 5;
+    let mut w: u64 = 2;
+
+    while i.pow(2) <= n {
+        if n%i == 0 {
+            return false;
+        }
+        i += w;
+        w = 6-w;
+    }
+
+    true
+}
+
 fn backwards_prime(start: u64, stop: u64) -> Vec<u64> {
-    // your code
     vec![start, stop]
 }
+
 fn testing(start: u64, stop: u64, exp: Vec<u64>) -> () {
     assert_eq!(backwards_prime(start, stop), exp)
 }