+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ---
 
+#### 19.09.2017
+
+* [C] Backwards Read Primes (CodeWars)
+
 #### 18.09.2017
 
 * [Rust] Backwards Read Primes (CodeWars)